Remove dev deploy split

+++ util/misc.sh
@@ -22,38 +22,17 @@ strip_files() {
 	done
 }
 
-split_dev_deploy() {
-	local lib f
-
-	if [ -d "$1/include" ]; then
-		mv "$1/include" "$2"
-	fi
-
-	if [ -d "$1/lib/pkgconfig" ]; then
-		mkdir -p "$2/lib/pkgconfig"
-		mv $1/lib/pkgconfig/* "$2/lib/pkgconfig"
-		rmdir "$1/lib/pkgconfig"
-	fi
+deploy_dev_cleanup() {
+	local f
 
 	if [ -d "$1/share/pkgconfig" ]; then
-		mkdir -p "$2/lib/pkgconfig"
-		mv $1/share/pkgconfig/* "$2/lib/pkgconfig"
+		mkdir -p "$1/lib/pkgconfig"
+		mv $1/share/pkgconfig/* "$1/lib/pkgconfig"
 		rmdir "$1/share/pkgconfig"
 	fi
 
 	for f in ${1}/lib/*.la; do
-		if [ -e "$f" ]; then
-			rm "$f"
-		fi
-	done
-
-	for f in ${1}/lib/*.a; do
-		if [ -f "$f" ]; then
-			mkdir -p "$2/lib"
-			mv ${1}/lib/*.a "$2/lib"
-		fi
-
-		break
+		[ ! -e "$f" ] || rm "$f"
 	done
 }
